+++ codex/serializers/fields/browser.py
@@ -5,9 +5,7 @@ from abc import ABC
 import pycountry
 from loguru import logger
 from pycountry.db import Database
-from rest_framework.serializers import (
-    ListField,
-)
+from rest_framework.serializers import ListField
 from typing_extensions import override
 
 from codex.choices.browser import (