+++ NEWS.md
@@ -29,6 +29,7 @@ border-radius: 128px;
 - Fix
   - Reading Order in reader was not always defaulting to your last Browser Top
     Group.
+  - Fixed filtering on Critical Rating and FileType
   - Bug displaying Age Rating, Original Format, Scan Info, Tagger in metadata.
   - Fixed database locking crashes.